test058: Fix CMakeLists on linux
CMake >= 3.11 requires CMP0072 to be configured on linux. It would
be fixed by setting `cmake_policy(SET CMP0072 NEW)`. I choose to
group-set policies by CMake version as it is encouraged in the
documentation:

https://cmake.org/cmake/help/latest/command/cmake_policy.html#setting-policies-by-cmake-version

The documentation says

    Note that the cmake_minimum_required(VERSION) command implicitly
    calls cmake_policy(VERSION) too.

but in reality `cmake` command complains if `cmake_policy` is not
called explicitly.