Add test217: Ansible multi-document YAML
We need to use a private function to load multi-document YAML using
Ansible's DataLoader. Of course we can load a YAML file and parse it
without DataLoader but then we are unable to load encrypted files.
And, decrypting arbitrary data requires a private API.

Maybe we should not try to load multi-document YAML in an Ansible
plugin.